chore(ui): rename DismissibleCard folder to dismissible-card
